requirements.txt
django
gunicorn
python-decouple
psycopg2-binary
djangorestframework
markdown
django-filter